refactor: Line 도메인 관련 피드백 반영 리팩터링(생성자 정리)
