+++ src/main/java/subway/line/domain/Line.java
@@ -20,36 +20,22 @@ public class Line {
     }
 
     public Line(Long id, String name, String color) {
+        this(name, color);
         this.id = id;
-        this.name = name;
-        this.color = color;
-    }
-
-    public Line(Long id, String name, String color, Sections sections) {
-        this.id = id;
-        this.name = name;
-        this.color = color;
-        this.sections = sections;
     }
 
     public Line(String name, String color, int extraFare) {
-        this.name = name;
-        this.color = color;
+        this(name, color);
         this.extraFare = extraFare;
     }
 
     public Line(Long id, String name, String color, int extraFare) {
+        this(name, color, extraFare);
         this.id = id;
-        this.name = name;
-        this.color = color;
-        this.extraFare = extraFare;
     }
 
     public Line(Long id, String name, String color, int extraFare, Sections sections) {
-        this.id = id;
-        this.name = name;
-        this.color = color;
-        this.extraFare = extraFare;
+        this(id, name, color, extraFare);
         this.sections = sections;
     }
