+++ src/main/java/subway/line/ui/LineController.java
@@ -2,10 +2,10 @@ package subway.line.ui;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.line.application.LineService;
 import subway.line.dto.LineRequest;
 import subway.line.dto.LineResponse;
 import subway.line.dto.SectionRequest;
+import subway.line.service.LineService;
 
 import java.net.URI;
 import java.sql.SQLException;
